requirements.txt
discord.py
paho-mqtt
python-dotenv